Visualise

Description

Provides two graph drawing functions - drawGraph and drawGraph' - that can invoke various graph drawing functions depending on their Method parameter. drawGraph draws a graph with default settings whereas drawGraph' can be provided with a Settings parameter. The Graph to be drawn is an instance of the type Graph as defined by Algebra.Graph.

The five graph drawing Methods defined for Method can be used for both drawing functions, apart from ExpressionTree which can only be used for drawGraph not drawGraph' due to its advanced parameters. Therefore drawExpressionTree and drawExpressionTree' are reexported from Visualise.ExpressionTree.

To draw a graph using Data.GraphViz the function drawWithGraphViz from Visualise.GraphViz can be used.

The constructors for the graph-drawing methods.

data Method

The graph drawing method

Constructors

Tree
The drawTree function from Visualise.Tree module will be used to draw the graph.

TreePartialOrder
The drawTreePartialOrder function Visualise.Tree module will be used to draw the graph, with all indirect dependencies removed by using the Coffman-Graham algorithm.

Circle
The Visualise.FlatCircle module will be used to draw the graph, with the verticies being placed on the vertecies of a regular polygon of n sides (where n is the number of graph vertices).

Hierarchical
The Visualise.Hierarchical module will be used to draw the graph as a hierarchical graph by grouping together vertices and groups of vertices with common connections.

Adaptive
The Visualise.FlatAdaptive module will be used to draw the graph, with the layout of vertices being dynamic depending on the connections between them.

ExpressionTree
The Visualise.ExpressionTree module will be used to draw an expression tree representation of the graph.

The two main graph drawing functions.

drawGraph :: (Show a, Eq a, Countable a) => Method -> Graph a -> Diagram B

Draw a graph using the specified method, uses default settings. To draw a graph using Visualise.GraphViz or Visualise.ExpressionTree, their own drawing functions must be used: for Visualise.GraphViz the function drawWithGraphViz and for Visualise.ExpressionTree the function drawExpressionTree or drawExpressionTree'.

drawGraph' :: (Show a, Eq a, Countable a) => Method -> (Graph a -> Settings) -> (a -> Diagram B) -> Graph a -> Diagram B

Draw a graph using the specified method, but also using the specified settings (provided as a function that will take a graph and return a Settings instance, see Visualise).

saveSVG :: FilePath -> Dimensions -> Diagram B -> IO ()

Saves a diagram to an SVG file at the specified FilePath, with the specified Dimensions. Only one dimension is needed, they are given in the format: (Maybe Double, Maybe Double)

The function for drawing with GraphViz, reexported from Visualise.GraphViz.

drawWithGraphViz

Arguments

:: (Ord a, Draw a)

=> GraphvizCommand
The way that Data.GraphViz should draw the graph, see Data.GraphViz.Commands.

-> Directed
Whether the graph is Directed.

-> Graph a
The graph that should be drawn.

-> IO (Diagram B)
The resultant graph drawing in an IO wrapper.

Uses Diagrams.TwoD.GraphViz as an interface to the Data.GraphViz library to produce a Diagrams representation of the provided graph using the provided GraphvizCommand. The resultant Diagram is wrapped in an IO from System.IO so must be bound to a function such as saveSVG to write it to a file. Uses the arrow paths produced by GraphViz but uses the custom functon draw to choose how to draw nodes. To draw graphs of graphs the Visualise.Tree module is used for the subgraph layouts.

The functions for drawing an expression tree from a graph reexported from Visualise.ExpressionTree.

drawExpressionTree :: (Show a, Countable a) => Graph a -> Diagram B

Visualises the provided Graph in the form of an expression tree with default nodes and link nodes, giving a Diagram as its output.

drawExpressionTree'

Arguments

:: (Show a, Countable a)

=> Diagram B
The default Empty node Diagram.

-> (a -> Diagram B)
A function which takes a vertex and produces a node Diagram.

-> Diagram B
A Diagram used for linking two expression tree branches when Overlay is used, by default defaultOverlayDiagram is used (a circle containing a +).

-> Diagram B
A Diagram used for linking two expression tree branches when Connect is used, by default defaultConnectDiagram is used (a circle containing a *).

-> Graph a
The Graph to be visualised as an expression tree.

-> Diagram B
The resultant expression tree Diagram.

Visualises the provided Graph in the form of an expression tree with custom nodes and link nodes. Folds through the graph with tuples of the type ('Diagram B', String) where the Diagram is the layout of the graph section and the Sting is an accumulator used for link identification when joining branches.