Visualise.Common

Description

Provides common types, typeclasses and functions used by the various drawing modules.

The Settings type used by most drawing modules to customise the drawing of a graph.

data Settings a

The Settings data type contains various fields for the graph drawing methods, some common between them and some specialised to a single method. Specialised fields use Maybe. The Lens library is used to enable easy access to and modification of these fields using getters and setters. To access a field the generated Lens getter functions can be used in conjunction with the inline get function (^.). For example the following test function gets the _directed field:

    getDirectedTest :: Setting a -> Directed
    getDirectedTest s = s ^. directed

And to set a value in an instance of Settings, Lens' setter functions can be used in combination with the inline set function (.~) as shown in the following test function: (Where (&) is a reverse application operator as defined by Lens and with is a synonym for def to get the default Settings value)

    setSettingsTest :: Setting a
    setSettingsTest = with & directed .~ Directed
                           & initPos .~ Just 2

This function uses a default Settings instance and sets the values of the _directed and _initPos fields. To set the node-drawing function to a custom function, the -XScopedTypeVariables language extension must be used due to Haskell's polymorphism. For example to use the drawDefaultHierNode function as the drawing function, the following function could be used:

    testSetNodeDrawFunction :: (Countable a, Show a) => Graph a -> Settings a
    testSetNodeDrawFunction g = (with :: (Show a) => Settings a) & (nodeDrawFunction .~ drawDefaultHierNode :: (Show a) => Settings a -> Settings a)

If the type definitions are not present there will be ambiguous types.

Each standard graph-drawing function in the library provides at least one default settings function providing a recommended set of default Settings to be used.

Constructors

Settings

Fields

_nodeDrawFunction :: a -> Diagram B

_dynamicHead :: Measure Double

The size of the arrow heads for the graph drawing, useful to scale with graph size. The default implementation for each of the drawing methods is similar, being based on: dynamicStyle small $ count g With normal being used in place of small for some methods.

_dynamicThick :: Measure Double

The thickness of the arrow shafts for the graph drawing, again useful to scale like dynamicHead like so: dynamicStyle thin $ count g

_directed :: Directed

Whether the graph is Directed (arrows on the connections between nodes). The type Directed is a boolean value that can either be Directed or Undirected.

_horizontalOrientation :: Maybe Bool

The orientation of the graph when using the Visualise.Tree module's functions.

_layerSpacing :: Maybe Double

The spacing between layers when using the Visualise.Tree module's functions.

_nodeSpacing :: Maybe Double

The spacing between vertices when using the Visualise.Tree module's functions.

_graphPadding :: Maybe Double

The padding around the graph when using the Visualise.Tree module's functions.

_colF :: Maybe (Int -> Colour Double)

The background colour of groups when using the Visualise.Hierarchical module.

_bgOp :: Maybe Double

The background opacity of groups when using the Visualise.Hierarchical module.

_initPos :: Maybe Int

The mode of initial positioning for the vertices when using Visualise.FlatAdaptive, changing this results in differently layed out graphs. Useful to experiment with to best display a graph.

The Node type which is used by the drawing modules to represent a vertex, containing a String representation and a Diagram representation.

data Node

The Node data type represents a node/vertex on a graph, storing the node's String name and its corrisponding diagram from the Diagrams library.

Constructors

Node

Fields

name :: String

Use name to get the String representation of a node.

diag :: Diagram B

Use diag to get the Diagrams representation of a node.

The ProcessedGraph type which is used by getVertices to represent a graph that has been folded through, with a list of vertices and a list of conenctions.

data ProcessedGraph

The ProcessedGraph data type contains a list of a graph's nodes along with the connections between them in a list of tuples.

Constructors

ProcessedGraph [Node] [(Node, Node)]

The Boolean type Directed which determines whether a graph is directed/has arrows on connections.

data Directed

The Directed data type is a boolean type which is used to determine if a graph is a directed or undiercted graph.

Constructors

Directed

Undirected

An array of tuples representing connections in a graph.

type ConnectList a = [(a, [a])]

The type ConnectList is an alias for an array of tuples that can be used where the first element in each tuple is a vertex and the second element is an array of the connected vertices - i.e. an adjacency list if the connected bvertices are only vertices dependant on the vertex.

The Draw TypeClass which determines how to draw a Node using the function draw.

class Draw a where

The TypeClass Draw is used to determine whether a value can be drawn, this is used for vertices as these must be able to be drawn.

Minimal complete definition

draw

Methods

draw :: a -> Diagram B

The draw function produces a Diagram from the drawable value.

draw :: Draw a => a -> Diagram B

The draw function produces a Diagram from the drawable value.

The Countable TypeClass which determines how to count the number of vertices in a graph/graph of graphs using the function count.

class Countable a where

The TypeClass Countable is used to determine if a type can be counted when counting the number of vertices in a graph/more specificaly a nested graph.

Minimal complete definition

count

Methods

count :: a -> Int

The count function gives an Integer representing the number of vertices in a.

count :: Countable a => a -> Int

The count function gives an Integer representing the number of vertices in a.

The default vertex-drawing functions.

drawDefaultNode :: Show a => a -> Diagram B

Produces a default node á la drawDefaultNodeWithSize with a size of '0.1'.

drawDefaultNodeWithSize :: Show a => Double -> a -> Diagram B

The default drawing function to produce a single node but of the specified size, in the form of a 'Diagram B', containing some showable text. Also makes the node a link, so if the drawn graph is saved as an SVG file each vertex is clickable with a JavaScript alert giving the vertex's name.

drawDefaultEmptyNode :: Double -> Diagram B

The default circle Diagram for an empty node of the size provided.

drawNodeWithEmptyFlag :: Show a => (a, Bool) -> Diagram B

The same as drawDefaultNode but instead takes a tuple of a showable value and a boolean flag. If the flag is True then it means the vertex should be displayed as an empty vertex with no contents, but if the flag is true the vertex is drawn as normal.

Gets a ProcessedGraph containing a graph's vertices and connections, a wrapper for namesAndConnections.

getVertices

Arguments

:: (Eq a, Show a)

=> (a -> Diagram B)
A function that takes a vertex of the graph of type a and produces a Diagram from it.

-> Graph a
The graph to be folded through.

-> ProcessedGraph
Contains the produced Node's and connections.

Essentially a wrapper for the recursive namesAndConnections function. Takes a Diagram-producing function and a Graph of the same type the Diagram function takes and then produces a ProcessedGraph containign all the vertices (their names and diagrams) and connections of the original graph.

Gets a ProcessedGraph containing a graph's vertices and connections, recursive requiring a counter string.

namesAndConnections

Arguments

:: (Eq a, Show a)

=> (a -> Diagram B)
A function that takes a vertex of the graph of type a and produces a Diagram from it.

-> Graph a
The graph to be folded through.

-> String
An accumulator used to produce identifiers for Empty vertices, each iteration an l or r is prepended.

-> ProcessedGraph
Contains the produced Node's and connections.

Recursively goes through the provided graph. Produces a list of Node's (so the name and associated Diagram for each vertex) and a list of connections between them.

Gets the vertices connected from and to a specified vertex.

connectedFrom :: Eq a => [(a, a)] -> ConnectList a

Similar to connectedTo, takes a list of connections and produces a ConnectList containing the vertices that each vertex depends on.

connectedTo

Arguments

:: Eq a

=> [(a, a)]
A list of tuples where the first tuple element and second tuple element are used to corrispond to the vertex at the tail of the conenction and at the head of connection respectively.

-> ConnectList a
The corrisponding adjacency list to the list of the connections.

Takes a list of connections and produces a corrisponding adjacency list (so tuples of a vertex and a list of vertices dependant on it) by using folds and recursion.

Used to adapt graph-drawing parameters in accordance with the size of a graph.

dynamicStyle :: Measure Double -> Int -> Measure Double

Used to produce a measurement for the size of Diagram properties from a default measurement and the size of a graph as an Integer. Multiplies the default value by 10 and divides this new value by the size of the graph.

The Lens setter functions to modify their corresponding Settings fields.

nodeDrawFunction :: forall a a. Lens (Settings a) (Settings a) (a -> Diagram B) (a -> Diagram B)

dynamicHead :: forall a. Lens' (Settings a) (Measure Double)

dynamicThick :: forall a. Lens' (Settings a) (Measure Double)

directed :: forall a. Lens' (Settings a) Directed

horizontalOrientation :: forall a. Lens' (Settings a) (Maybe Bool)

layerSpacing :: forall a. Lens' (Settings a) (Maybe Double)

nodeSpacing :: forall a. Lens' (Settings a) (Maybe Double)

graphPadding :: forall a. Lens' (Settings a) (Maybe Double)

colF :: forall a. Lens' (Settings a) (Maybe (Int -> Colour Double))

bgOp :: forall a. Lens' (Settings a) (Maybe Double)

initPos :: forall a. Lens' (Settings a) (Maybe Int)