Visualise.FlatAdaptive

Description

Draws a graph in an adaptive way by grouping together connected vertices. Works with some graphs however currently needs work.

Provides two drawing functions: drawFlatAdaptive which uses default Settings and drawFlatAdaptive' which can be customised.

Draws a graph with an adaptive layout with default Settings.

drawFlatAdaptive :: (Show a, Eq a, Countable a) => Graph a -> Diagram B

Uses drawFlatAdaptive' with default Settings provided by defaultAdaptiveSettings.

Draws a graph with an adaptive layout with customised Settings.

drawFlatAdaptive' :: (Show a, Eq a) => Settings -> (a -> Diagram B) -> Graph a -> Diagram B

Draws a flat graph with an adaptive layout.

The default Settings function used by drawFlatAdaptive.

defaultAdaptiveSettings :: Countable a => Graph a -> Settings

The default Settings function for graphs generated with Visualise.FlatAdaptive.