Visualise.FlatCircle

Description

Draws a Graph with each vertex at a vertex of a regular polygon with the same number of sides as the Graph does vertices.

Provides a function which draws the Graph with default Settings, drawFlatCircle, and a second function which draws the graph with custom Settings.

Draws a Graph with default Settings.

drawFlatCircle :: (Show a, Eq a, Countable a) => Graph a -> Diagram B

Draws the provided graph as a circle with default settings.

Draws a Graph with customised Settings.

drawFlatCircle' :: (Show a, Eq a, Countable a) => Settings -> (a -> Diagram B) -> Graph a -> Diagram B

Draws the provided graph with each graph vertex at a vertex of a regular polygon with the same number of sides