Visualise.GraphViz

Documentation

drawWithGraphViz

Arguments

:: (Ord a, Draw a)

=> GraphvizCommand
The way that Data.GraphViz should draw the graph, see Data.GraphViz.Commands.

-> Directed
Whether the graph is Directed.

-> Graph a
The graph that should be drawn.

-> IO (Diagram B)
The resultant graph drawing in an IO wrapper.

Uses Diagrams.TwoD.GraphViz as an interface to the Data.GraphViz library to produce a Diagrams representation of the provided graph using the provided GraphvizCommand. The resultant Diagram is wrapped in an IO from System.IO so must be bound to a function such as saveSVG to write it to a file.

Orphan instances

(Ord a, Show a) => Ord (Graph a)

Methods

compare :: Graph a -> Graph a -> Ordering

(<) :: Graph a -> Graph a -> Bool

(<=) :: Graph a -> Graph a -> Bool

(>) :: Graph a -> Graph a -> Bool

(>=) :: Graph a -> Graph a -> Bool

max :: Graph a -> Graph a -> Graph a

min :: Graph a -> Graph a -> Graph a