Visualise.Hierarchical

Description

Draws a graph as a hierarchical graph by grouping together vertices with common connections and containing them with a box.

The main drawing function drawHier draws a hierarchical graph with default Settings as defined by defaultHierSettings with alternating background colours as defined by alternatingColour, as well as a second function drawHier' which produces a customised drawing.

The hierarchical graph-drawing functions.

drawHier :: (Show a, Countable a) => Graph a -> Diagram B

The default drawing function which uses drawHier' with the default Settings provided by the function defaultHierSettings. The background colour for each layer alternates using the alternatingColour function and have the opacity of '1'.

drawHier' :: (Show a, Countable a) => Settings a -> Graph a -> Diagram B

Takes Settings, a vertex-to-Diagram function and a graph to produce a hierarchical representation for the graph in the form of a Diagram.

The default Settings used by drawHier.

defaultHierSettings :: (Countable a, Show a) => Graph a -> Settings a

Produces the default Settings for the supplied graph. The arrow heads and shafts automatically scale to the number of graph vertices, the graph is Directed and the group background colours alternate using alternatingColour by default with an opacity of 1.

The default node-drawing function used by defaultHierSettings.

drawDefaultHierNode :: Show a => a -> Diagram B

The default function to produce a Diagram for a vertex. Calls show on the provided value and puts the text in a circle.

The alternating-colour-producing function using by defaultHierSettings.

alternatingColour

Arguments

:: Int
The depth of the current layer of the graph.

-> Colour Double
The background colour for the layer.

If the current depth is odd the colour is red and if it's even then the colour is cyan.