Visualise.Tree

Documentation

drawTree :: (Show a, Eq a, Countable a) => Graph a -> Diagram B

Draws the provided graph as a tree, producing a Diagram using drawTree', using the default settings.

drawTree' :: (Show a, Eq a, Countable a) => (Graph a -> Settings) -> (a -> Diagram B) -> Graph a -> Diagram B

Draws the graph provided with visualiseTree, using the provided function to produce the visualisation Settings. For more information on customisable settings see drawTree'. Uses getVertices to produce a ProcessedGraph with the graph's verices and conenctions, then uses visualiseTree with these.

drawTreePartialOrder :: (Show a, Eq a, Countable a) => Graph a -> Diagram B

Removes indirect connections from the graph and produces a Diagram, using drawTreePartialOrder' with the default drawing Settings provided defaultTreeSettings.

drawTreePartialOrder' :: (Show a, Eq a, Countable a) => (Graph a -> Settings) -> (a -> Diagram B) -> Graph a -> Diagram B

Removes indirect connections from the graph and draws it using visualiseTree, producing a Diagram. Provides a parameter to supply a function that takes the graph and returns an instance of the Settings type that will give the drawing functions the graph visualisation settings. An example of such a function is the defaultTreeSettings function: defaultTreeSettings :: (Countable a) => Graph a -> Settings defaultTreeSettings g = Settings (dynamicStyle small $ count g) (dynamicStyle thin $ count g) Directed (Just 0.2) (Just 0.3) (Just 0.1) Nothing Nothing Nothing This shows that the arrow head size, arrow shaft thickness, whether the graph is directed, the horizontal and vertical spacing between vertices and the graph frame padding can be customised.

The indirect conenctions are removed by using reduction and the graph is processed by getVertices to produce a ProcessedGraph instance containing the vertices and their connections.