Visualise

Documentation

data Method

The graph drawing method

Constructors

Tree
The drawTree function from Visualise.Tree module will be used to draw the graph

TreePartialOrder
The drawTreePartialOrder function Visualise.Tree module will be used to draw the graph, with all indirect dependencies removed by using the Coffman-Graham algorithm

Circle
The Visualise.FlatCircle module will be used to draw the graph, with the verticies being placed on the vertecies of a regular polygon of n sides (where n is the number of graph vertices)

Hierarchical
The Visualise.Hierarchical module will be used to draw the graph as a hierarchical graph by grouping together vertices and groups of vertices with common connections

Adaptive
The Visualise.FlatAdaptive module will be used to draw the graph, with the layout of vertices being dynamic depending on the connections between them.

drawGraph :: (Show a, Eq a, Countable a) => Method -> Graph a -> Diagram B

Draw a graph using the specified method, uses default settings

drawGraph' :: (Show a, Eq a, Countable a) => Method -> (Graph a -> Settings) -> (a -> Diagram B) -> Graph a -> Diagram B

Draw a graph using the specified method, but also using the specified settings (provided as a function that will take a graph and return a Settings instance, see Visualise)

saveSVG :: FilePath -> Dimensions -> Diagram B -> IO ()

Saves a diagram to an SVG file at the specified FilePath, with the specified Dimensions. Only one dimension is needed, they are given in the format: (Maybe Double, Maybe Double)