Visualise.Common

Documentation

data Settings

The Settings data type contains various fields for the graph drawing methods, some common between them and some specialised to a single method. To give custom settings to the drawGraph' function, a function which takes a graph as an argument and returns a Settings instance will be given to it.

Constructors

Settings

Fields

dynamicHead :: Measure Double

The size of the arrow heads for the graph drawing, useful to scale with graph size. The default implementation for each of the drawing methods is similar, being based on: dynamicStyle small $ count g With normal being used in place of small for some methods

dynamicThick :: Measure Double

The thickness of the arrow shafts for the graph drawing, again useful to scale like dynamicHead like so: dynamicStyle thin $ count g

directed :: Directed

Whether the graph is Directed (arrows on the connections between nodes). The type Directed is a boolean value that can either be Directed or Undirected

layerSpacing :: Maybe Double

The spacing between layers when using the Visualise.Tree module's functions

nodeSpacing :: Maybe Double

The spacing between vertices when using the Visualise.Tree module's functions

graphPadding :: Maybe Double

The padding around the graph when using the Visualise.Tree module's functions

colF :: Maybe (Int -> Colour Double)

The background colour of groups when using the Visualise.Hierarchical module

bgOp :: Maybe Double

The background opacity of groups when using the Visualise.Hierarchical module

initPos :: Maybe Int

The mode of initial positioning for the vertices when using Visualise.FlatAdaptive, changing this results in differently layed out graphs. Useful to experiment with to best display a graph.

data Node

The Node data type represents a node/Vertex on a graph, storing the node's String name and its corrisponding diagram from the Diagrams library

Constructors

Node

Fields

name :: String

Use name to get the String representation of a node

diag :: Diagram B

Use diag to get the Diagrams representation of a node

data ProcessedGraph

The ProcessedGraph data type contains a list of a Graph's nodes along with the connections between them in a list of tuples.

Constructors

ProcessedGraph [Node] [(Node, Node)]

data Directed

The Directed data type is a boolean type which is used to determine if a graph is a directed or undiercted graph.

Constructors

Directed

Undirected

type ConnectList a = [(a, [a])]

The type ConnectList is an alias for an array of tuples that can be used where the first element in each tuple is a vertex and the second element is an array of the connected vertices - i.e. an adjacency list if the connected bvertices are only vertices dependant on the vertex.

class Draw a where

The TypeClass Draw is used to determine whether a value can be drawn, this is used for vertices as these must be able to be drawn.

Minimal complete definition

draw

Methods

draw :: a -> Diagram B

The draw function produces a Diagram from the drawable value.

class Countable a where

The TypeClass Countable is used to determine if a type can be counted when counting the number of vertices in a graph/more specificaly a nested graph.

Minimal complete definition

count

Methods

count :: a -> Int

The count function gives an Integer representing the number of vertices in a.

draw :: Draw a => a -> Diagram B

The draw function produces a Diagram from the drawable value.

count :: Countable a => a -> Int

The count function gives an Integer representing the number of vertices in a.

drawDefaultNode :: Show a => a -> Diagram B

The default drawing function to produce a single node, in the form of a 'Diagram B', containing some showable text. Also makes the node a link, so if the drawn graph is saved as an SVG file each vertex is clickable with a JavaScript alert giving the vertex's name.

drawNodeWithEmptyFlag :: Show a => (a, Bool) -> Diagram B

The same as drawDefaultNode but instead takes a tuple of a showable value and a boolean flag. If the flag is True then it means the vertex should be displayed as an empty vertex with no contents, but if the flag is true the vertex is drawn as normal.

getVertices

Arguments

:: (Eq a, Show a)

=> (a -> Diagram B)
A function that takes a vertex of the Graph of type a and produces a Diagram from it.

-> Graph a
The Graph to be folded through.

-> ProcessedGraph
Contains the produced Node's and connections.

Essentially a wrapper for the recursive namesAndConnections function. Takes a Diagram-producing function and a Graph of the same type the Diagram function takes and then produces a ProcessedGraph containign all the vertices (their names and diagrams) and connections of the original graph.

connectedFrom :: Eq a => [(a, a)] -> ConnectList a

Similar to connectedTo, takes a list of connections and produces a ConnectList containing the vertices that each vertex depends on.

connectedTo

Arguments

:: Eq a

=> [(a, a)]
A list of tuples where the first tuple element and second tuple element are used to corrispond to the Vertex at the tail of the conenction and at the head of connection respectively.

-> ConnectList a
The corrisponding adjacency list to the list of the connections.

Takes a list of connections and produces a corrisponding adjacency list (so tuples of a vertex and a list of vertices dependant on it) by using folds and recursion.

dynamicStyle :: Measure Double -> Int -> Measure Double

Used to produce a measurement for the size of Diagram properties from a default measurement and the size of a graph as an Integer. Multiplies the default value by 10 and divides this new value by the size of the graph.