Visualise.ExpressionTree

Description

Visualises an algebraic Graph expression as an expression tree: each Vertex is a leaf and are joined by a node which is represented by a different Diagram depending on if the two vertices are joined by a Connect or an Overlay. By default Connect is represented by a * in a circle and Overlay is represented by a + in a circle.

drawExpressionTree can be used to draw the Graph representation with default Settings and Drawings. This function can be invoked by the main drawing function drawGraph in the main Visualise module, using the ExpressionTree Method constructor. drawExpressionTree' can also be used but as well as a Graph extra parameters can be provided to determine how to draw different nodes in the expression tree, however this function cannot be invoked using drawGraph' from Visualise so must be used directly.

Visualises a Graph as an expression tree with default Settings.

drawExpressionTree :: (Show a, Countable a) => Graph a -> Diagram B

Visualises the provided Graph in the form of an expression tree with default nodes and link nodes, giving a Diagram as its output.

Visualises a Graph as an expression tree drawn using the specified parameters.

drawExpressionTree'

Arguments

:: (Show a, Countable a)

=> Diagram B
The default Empty node Diagram.

-> (a -> Diagram B)
A function which takes a vertex and produces a node Diagram.

-> Diagram B
A Diagram used for linking two expression tree branches when Overlay is used, by default defaultOverlayDiagram is used (a circle containing a +).

-> Diagram B
A Diagram used for linking two expression tree branches when Connect is used, by default defaultConnectDiagram is used (a circle containing a *).

-> Graph a
The Graph to be visualised as an expression tree.

-> Diagram B
The resultant expression tree Diagram.

Visualises the provided Graph in the form of an expression tree with custom nodes and link nodes. Folds through the graph with tuples of the type ('Diagram B', String) where the Diagram is the layout of the graph section and the Sting is an accumulator used for link identification when joining branches.