Visualise.ExpressionTree

Documentation

drawExpressionTree :: (Show a, Countable a) => Graph a -> Diagram B

Visualises the provided Graph in the form of an expression tree with default nodes and link nodes, giving a Diagram as its output.

drawExpressionTree'

Arguments

:: (Show a, Countable a)

=> Diagram B
The default Empty node Diagram.

-> (a -> Diagram B)
A function which takes a vertex and produces a node Diagram.

-> Diagram B
A Diagram used for linking two expression tree branches when Overlay is used, by default defaultOverlayDiagram is used (a circle containing a +).

-> Diagram B
A Diagram used for linking two expression tree branches when Connect is used, by default defaultConnectDiagram is used (a circle containing a *).

-> Graph a
The Graph to be visualised as an expression tree.

-> Diagram B
The resultant expression tree Diagram.

Visualises the provided Graph in the form of an expression tree with custom nodes and link nodes. Folds through the graph