Visualise.FlatCircle

Documentation

data Settings

The Settings data type contains various fields for the graph drawing methods, some common between them and some specialised to a single method. To give custom settings to the drawGraph' function, a function which takes a graph as an argument and returns a Settings instance will be given to it.

Constructors

Settings

Fields

dynamicHead :: Measure Double

The size of the arrow heads for the graph drawing, useful to scale with graph size. The default implementation for each of the drawing methods is similar, being based on: dynamicStyle small $ count g With normal being used in place of small for some methods

dynamicThick :: Measure Double

The thickness of the arrow shafts for the graph drawing, again useful to scale like dynamicHead like so: dynamicStyle thin $ count g

directed :: Directed

Whether the graph is Directed (arrows on the connections between nodes). The type Directed is a boolean value that can either be Directed or Undirected

layerSpacing :: Maybe Double

The spacing between layers when using the Visualise.Tree module's functions

nodeSpacing :: Maybe Double

The spacing between vertices when using the Visualise.Tree module's functions

graphPadding :: Maybe Double

The padding around the graph when using the Visualise.Tree module's functions

colF :: Maybe (Int -> Colour Double)

The background colour of groups when using the Visualise.Hierarchical module

bgOp :: Maybe Double

The background opacity of groups when using the Visualise.Hierarchical module

initPos :: Maybe Int

The mode of initial positioning for the vertices when using Visualise.FlatAdaptive, changing this results in differently layed out graphs. Useful to experiment with to best display a graph.

drawFlatCircle :: (Show a, Eq a, Countable a) => Graph a -> Diagram B

Draws the provided graph as a circle with default settings.

drawFlatCircle' :: (Show a, Eq a, Countable a) => (Graph a -> Settings) -> (a -> Diagram B) -> Graph a -> Diagram B