Visualise.FlatCircle

Description

Draws a graph with each vertex at a vertex of a regular polygon with the same number of sides as the graph does vertices.

Provides a function which draws the graph with default Settings, drawFlatCircle, and a second function which draws the graph with custom Settings.

Draws a graph with default Settings.

drawFlatCircle :: (Show a, Eq a, Countable a) => Graph a -> Diagram B

Draws the provided graph as a circle with default settings.

Draws a graph with customised Settings.

drawFlatCircle' :: (Show a, Eq a, Countable a) => Settings a -> Graph a -> Diagram B

Draws the provided graph with each graph vertex at a vertex of a regular polygon with the same number of sides