Visualise.GraphViz

Description

Uses Diagrams.TwoD.GraphViz to enable Data.GraphViz to be used in combination with Diagrams to draw a Graph layed out with a specified GraphvizCommand with drawWithGraphViz.

The TypeClass that determines how to draw Graph vertices, exported to enable Graphs of custom types.

class Draw a

Used to determine how to draw a vertex, different to Visualise.Common.Draw.

Minimal complete definition

draw

The drawing function that can take a GraphvizCommand to visualise a Graph.

drawWithGraphViz

Arguments

:: (Ord a, Draw a)

=> GraphvizCommand
The way that Data.GraphViz should draw the graph, see Data.GraphViz.Commands.

-> Directed
Whether the graph is Directed.

-> Graph a
The graph that should be drawn.

-> IO (Diagram B)
The resultant graph drawing in an IO wrapper.

Uses Diagrams.TwoD.GraphViz as an interface to the Data.GraphViz library to produce a Diagrams representation of the provided graph using the provided GraphvizCommand. The resultant Diagram is wrapped in an IO from System.IO so must be bound to a function such as saveSVG to write it to a file. Uses the arrow paths produced by GraphViz but uses the custom functon draw to choose how to draw nodes. To draw graphs of graphs the Visualise.Tree module is used for the subgraph layouts.

Orphan instances

(Ord a, Show a) => Ord (Graph a)
Defines how to compare graphs.

Methods

compare :: Graph a -> Graph a -> Ordering

(<) :: Graph a -> Graph a -> Bool

(<=) :: Graph a -> Graph a -> Bool

(>) :: Graph a -> Graph a -> Bool

(>=) :: Graph a -> Graph a -> Bool

max :: Graph a -> Graph a -> Graph a

min :: Graph a -> Graph a -> Graph a