Visualise.Tree

Description

Enables a graph to be drawn in a Tree style. Provides four drawing functions: drawTree to draw the graph with default Settings and drawTree' allows the Settings to be customised, as well as drawTreePartialOrder which uses the Coffman-Graham algorithm to remove indirect connections for directed acyclic graphs and drawTreePartialOrder' which does the same but allows the Settings to be specified.

The main drawing function: the full graph as specified with default Settings.

drawTree :: (Show a, Eq a, Countable a) => Graph a -> Diagram B

Draws the provided graph as a tree, producing a Diagram using drawTree', using the default settings function defaultTreeSettings. Self-loops are supported.

Allows the Settings to be customised.

drawTree' :: (Show a, Eq a, Countable a) => Settings a -> Graph a -> Diagram B

Draws the graph provided with visualiseTree, using the provided visualisation Settings. Uses getVertices to produce a ProcessedGraph with the graph's verices and conenctions, then uses visualiseTree with these, self-loops are supported.

Removes indirect connections using the Coffman-Graham algorithm before drawing.

drawTreePartialOrder :: (Show a, Eq a, Countable a) => Graph a -> Diagram B

Removes indirect connections from the graph and produces a Diagram, using drawTreePartialOrder' with the default drawing Settings provided defaultTreeSettings. Self-loops are not supported.

drawTreePartialOrder' :: (Show a, Eq a, Countable a) => Settings a -> Graph a -> Diagram B

Removes indirect connections from the graph and draws it using visualiseTree, producing a Diagram. A Settings instance can be provided to customise the drawing of the graph. The indirect conenctions are removed by using reduction and the graph is processed by getVertices to produce a ProcessedGraph instance containing the vertices and their connections. Self-loops are not supported.

The default Settings functions used with drawTree' and drawTreePartialOrder'.

defaultTreeSettings :: (Countable a, Show a) => Graph a -> Settings a

Generates a default Settings instance from the provided graph. The arrow head size and shaft thickness vary in accordance with the graph size and the graph is Directedand vertically orientated with layer separation, vertex (horizonal) separation and frame padding of 0.2, 0.3 and 0.1 respectively.

defaultTreeSettingsHorizontal :: (Countable a, Show a) => Graph a -> Settings a

Generates a default Settings instance from the provided graph, but rotated so the graph is horizonal instead of vertical. The arrow head suze and shaft thickness vary in accordance with the graph size and the graph is Directed and horizontally orientated with layer separation, vertex (horizonal) separation and frame padding of 0.2, 0.3 and 0.1 respectively.